requirements.txt
GoogleAppEngineCloudStorageClient
python-gcm